added user registration
This commit introduces the user registration feature, including the registration form, API route, and related components. The form includes validation for email, password, and confirm password fields. The backend route for registration has been updated to remove the authJWT middleware temporarily. Additionally, the middleware has been updated to include the `/register` route as a public route. The login page now includes a link to the registration page, and vice versa.